+++ requirements-dev.txt
@@ -6,6 +6,7 @@ pytest-timeout>=1.2.0
 pytest-random>=0.02
 pytest-cov
 pytest-catchlog
+flaky
 hypothesis
 grequests
 psutil
